initial support for jumping into jars.
the file we want to jump to is extracted to /tmp/$basename (conflicts
possible) and that name is passed to the regular opener.

in the future it might be more interesting to support opening arbitrary
urls, e.g. easily open web pages, images, directories, zip-files/jars
and more. but that would require quite a lot of changes. (maybe ask
someone more experienced about this. possibly a prototype could be
developed as a plugin at first?)